# pre-compile regexp2 with regexp2cg