--- a/codec/cl100k_base.go
+++ b/codec/cl100k_base.go
@@ -4,10 +4,13 @@ import "github.com/dlclark/regexp2"
 
 func NewCl100kBase() *Codec {
 	cl100kBaseVocabOnce.Do(cl100kBaseVocabInit)
+
+	splitRegexp := regexp2.MustCompile(`(?i:'s|'t|'re|'ve|'m|'ll|'d)|[^\r\n\p{L}\p{N}]?\p{L}+|\p{N}{1,3}| ?[^\s\p{L}\p{N}]+[\r\n]*|\s*[\r\n]+|\s+(?!\S)|\s+`, regexp2.None)
+
 	return &Codec{
 		name:        "cl100k_base",
 		vocabulary:  cl100kBaseVocab,
-		splitRegexp: regexp2.MustCompile(`(?i:'s|'t|'re|'ve|'m|'ll|'d)|[^\r\n\p{L}\p{N}]?\p{L}+|\p{N}{1,3}| ?[^\s\p{L}\p{N}]+[\r\n]*|\s*[\r\n]+|\s+(?!\S)|\s+`, regexp2.None),
+		splitRegexp: splitRegexp,
 		specialTokens: map[string]uint{
 			"<|endoftext|>":   100257,
 			"<|fim_prefix|>":  100258,