--- a/codec/regexp.go
+++ b/codec/regexp.go
@@ -1,0 +1,2 @@
+//go:generate go tool github.com/dlclark/regexp2cg -package codec  -o regexp.gen.go
+package codec