--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,12 @@
 module github.com/tiktoken-go/tokenizer
 
-go 1.20
+go 1.21.4
 
-require github.com/dlclark/regexp2 v1.9.0
+toolchain go1.22.0
+
+require github.com/dlclark/regexp2 v1.11.5-0.20240806004527-5bbbed8ea10b
+
+require (
+	github.com/dlclark/regexp2cg v0.1.0 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+)