--- a/go.mod
+++ b/go.mod
@@ -4,9 +4,11 @@ go 1.21.4
 
 toolchain go1.22.0
 
+tool github.com/dlclark/regexp2cg
+
 require github.com/dlclark/regexp2 v1.11.5-0.20240806004527-5bbbed8ea10b
 
 require (
-	github.com/dlclark/regexp2cg v0.1.0 // indirect
+	github.com/dlclark/regexp2cg v0.2.0 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
 )