@bluescreen10 2024-12-04
I was looking for something like this a while ago. Will test and give you feedback. Thanks!