@bluescreen10 2024-12-05
Hey,

I can't get the regeneration to work, can you tell me what command did you execute to regenerate the code?

Ideally would like to make a comment in some file `//go:generate regexp2gc ...`

```
$ regexp2cg -path codec -package codec  -o codec/regexp.gen.go
2024/12/05 09:21:02 Create regexp for path ./tokenizer/codec, include tests=false
2024/12/05 09:21:02 file ./tokenizer/codec/r50k_base.go imports regexp2
2024/12/05 09:21:02 unknown ast node type for options: *ast.SelectorExpr &{X:regexp2 Sel:None}
2024/12/05 09:21:02 file./tokenizer/codec/cl100k_base.go imports regexp2
2024/12/05 09:21:02 unknown ast node type for options: *ast.SelectorExpr &{X:regexp2 Sel:None}
2024/12/05 09:21:02 file ./tokenizer/codec/codec.go imports regexp2
2024/12/05 09:21:02 file ./tokenizer/codec/regexp.gen.go imports regexp2
2024/12/05 09:21:02 file ./tokenizer/codec/o200k_base.go imports regexp2
2024/12/05 09:21:02 unknown ast node type for options: *ast.SelectorExpr &{X:regexp2 Sel:None}
2024/12/05 09:21:02 file ./tokenizer/codec/p50k_base.go imports regexp2
2024/12/05 09:21:02 unknown ast node type for options: *ast.SelectorExpr &{X:regexp2 Sel:None}
2024/12/05 09:21:02 file ./tokenizer/codec/p50k_edit.go imports regexp2
2024/12/05 09:21:02 unknown ast node type for options: *ast.SelectorExpr &{X:regexp2 Sel:None}
```