@terwey 2024-12-05
Please check my PR on the tools repo and you have to use my modifications on this repo to get it to work.

It doesn't work with MustCompile being inside a struct declaration.