@bluescreen10 2024-12-05
I checked out this pr so the regexps are on a separate variable. 

what's your PR on the tools repo?