@terwey 2024-12-05
https://github.com/dlclark/regexp2cg/pull/2

I just ran it inside `codec` after doing a `go build` from my tools PR, without any parameters.