@bluescreen10 2024-12-05
Ahh, got it! I'd like your PR on regexp2cg to be merged first before approving this PR.

Before switching the Regexp Engine I'd like to make sure this library will be maintained.

That said, I'll try to nudge your PR to dlclrak/regexp2.

Thanks for your patience.