@terwey 2024-12-05
Understandable! No rush, just nice to see performance improvements like this are possible.