@terwey 2025-01-30
@bluescreen10 just a small bump, https://github.com/dlclark/regexp2cg/pull/2 has been merged