@bluescreen10 2025-02-04
I unfortunately, even though it was merged, it wasn't released yet. I've created an issue to track that

https://github.com/dlclark/regexp2cg/issues/4