@geraldstanje1 2025-02-09
@bluescreen10 how to get it without official release? how to build the lib? does it also improve memory usage?